requirements.txt
lask 
transformers 
torch 
numpy<2 